split up some classes into their own files
